+++ app.js
@@ -1,15 +1,2 @@
-// Path module
+// file system module
 
-const path = require('path')
-const os = require('os')
-
-// A platform specific separator
-const platformSeparator = path.sep;
-
-console.log(`The separator for ${os.type()} OS is "${platformSeparator}"`);
- 
-const filePath = path.join('\content','subfolder','test.txt');
-console.log(`The required path is ${filePath}`);
-
-const base = path.basename(filePath);
-console.log(`The base filename is ${base}`);
